+++ main.gd
@@ -12,20 +12,20 @@ var pointer_scene: PackedScene = preload("res://hud/pointer.tscn")
 
 @export var MIN_SPAWN_INTERVAL := 0.8
 @export var MAX_SPAWN_INTERVAL := 1.0
-@export var SPAWN_INTERVAL_STEP := 0.1
+@export var SPAWN_INTERVAL_STEP := 0.05
 @export var SPAWN_INTERVAL_VARIANCE := 0.05
 
-@export var MIN_MOB_SPEED := 100.0
+@export var MIN_MOB_SPEED := 80.0
 @export var MAX_MOB_SPEED := 150.0
-@export var MOB_SPEED_STEP := 5
+@export var MOB_SPEED_STEP := 7.5
 
 @export var MIN_MOB_SCALE := 1.0
 @export var MAX_MOB_SCALE := 2.0
 @export var MOB_SCALE_STEP := 0.3
 
-@export var MIN_DIRECTION_VARIENCE := 4
+@export var MIN_DIRECTION_VARIENCE := 2
 @export var MAX_DIRECTION_VARIENCE := 36
-@export var DIRECTION_VARIENCE_STEP := 4
+@export var DIRECTION_VARIENCE_STEP := 2
 
 
 var score: int
@@ -51,6 +51,7 @@ func _ready() -> void:
 
 
 func game_over() -> void:
+	hud.add_difficulty_message("Better luck next time...")
 	game_over_music.play()
 	bg_music.volume_db = -28
 	if is_high_score(score):
@@ -60,6 +61,7 @@ func game_over() -> void:
 		hud.show_high_score(load_highscore())
 	$ScoreTimer.stop()
 	mob_timer.stop()
+	difficulty_timer.stop()
 	hud.show_game_over()
 
 
@@ -71,8 +73,12 @@ func new_game() -> void:
 	$StartTimer.start()
 	hud.update_score(score)
 	hud.hide_high_score()
-	hud.show_message("Get Ready")
+	hud.clear_difficulty_messages()
 	hud.restore_all_hearts()
+	hud.show_message("")
+	hud.add_difficulty_message("Get Ready...")
+	await hud.message_timer.timeout
+	hud.add_difficulty_message("ASTEROIDS INCOMING!!!")
 	bg_music.volume_db = -18
 	current_spawn_interval = FIRST_SPAWN_INTERVAL
 	current_speed = MIN_MOB_SPEED
@@ -156,21 +162,19 @@ func _on_difficulty_timer_timeout() -> void:
 		"interval":
 			current_spawn_interval -= SPAWN_INTERVAL_STEP
 			current_spawn_interval = clamp(current_spawn_interval, MIN_SPAWN_INTERVAL, MAX_SPAWN_INTERVAL)
-			# print("SPAWN INTERVAL DECREASED: ", current_spawn_interval)
-			# print("==============================")
+			hud.add_difficulty_message("Asteroids are now spawning even faster!")
 		"speed":
 			current_speed += MOB_SPEED_STEP
 			current_speed = clamp(current_speed, MIN_MOB_SPEED, MAX_MOB_SPEED)
-			# print("MOB SPEED INCREASED. Now: ", current_speed)
-			# print("==============================")
+			hud.add_difficulty_message("Asteroids are now even faster !")
 		"scale":
 			current_scale += MOB_SCALE_STEP
 			current_scale = clamp(current_scale, MIN_MOB_SCALE, MAX_MOB_SCALE)
-			# print("MOB SCALE INCREASED. Now: ", current_scale)
-			# print("==============================")
+			hud.add_difficulty_message("Asteroids are now even bigger!")
 		"aim":
 			current_direction_varience += DIRECTION_VARIENCE_STEP
 			current_direction_varience = clamp(current_direction_varience, MIN_DIRECTION_VARIENCE, MAX_DIRECTION_VARIENCE)
+			hud.add_difficulty_message("Asteroids are now aiming even better!")
 		"_":
 			pass
 	difficulty_timer.resetTimer()